NOT working copy, try to turn everything into unique_ptr, which need to define something like static_cast_unqiue_ptr. Let's redo it some other way

+++ src/algorithm/em_algorithm_binomial.cc
@@ -21,7 +21,7 @@ void EmAlgorithmBinomial::InitialiseSummaryStat() {
     em_stat_local = std::unique_ptr<EmSummaryStat>(new EmSummaryStatBinomial());
     em_stat_local->print();
 
-    for (int i = 0; i < num_category; ++i) {
+    for (size_t i = 0; i < num_category; ++i) {
         all_em_stats.emplace_back(new EmSummaryStatBinomial());
     }
